+++ vgo-core/src/test/kotlin/com/jzbrooks/vgo/core/transformation/BakeTransformationsTests.kt
@@ -1,4 +1,4 @@
-package com.jzbrooks.vgo.core.optimization
+package com.jzbrooks.vgo.core.transformation
 
 import assertk.all
 import assertk.assertThat
