+++ vgo-core/src/test/kotlin/com/jzbrooks/vgo/core/transformation/RemoveEmptyGroupsTests.kt
@@ -1,4 +1,4 @@
-package com.jzbrooks.vgo.core.optimization
+package com.jzbrooks.vgo.core.transformation
 
 import assertk.assertThat
 import assertk.assertions.hasSize
